Averaging Example
"attached a minimal, correct averaging example. This is for the data of Aurnhammer et al., (2021), but can be easily adjusted to the other data sets as well. For an ML project, you most likely want the by-trial averages (which also reduces the data set from ~1gb to ~1mb)" H.Brouwer